Align DEBT-364 continuity tests with repo harness
